Completely remove `:/` baseline skip
This rolls back all changes made to work around the Git 2.47.* bug
that underlies #1622, including the change made in the immediately
preceding commit.

This undoes the changes to `regex.rs` from #1635, #1719, #1774, and
2400158 (which is the immediately preceding commit). That file is
now as it was in 3745212.

The rationale is that the disadvantages of inverting the CI check
and extending the suppresson, as described in the previous commit,
really outweigh the advantages. This is mainly due to the risk of
generating archives that should not be committed but seem based on
test results like they could be committed.

(The suppressions being removed here will most likely not be needed
in the future, but if they are then this commit can be reverted,
and the suppression will be done locally but on on CI, consistently
across feature for which the relevant tests are run, and only when
Git is found to be a version in the 2.47.* range.)

Closes #1622