Suppress color in `cargo tree` warnings for simpler grepping
Without this, the `check` recipe in the `justfile` was failing on
CI due to the change in the previous commit. On CI, we set
`CARGO_TERM_COLOR` to `always`. The `--color` option takes
precedence over the `CARGO_TERM_COLOR` environment variable.

The suppresion is only done on those specific command whose
stderr is being parsed. Other colorization remains allowed,
including on CI.

(An alternative to using `--color` here could be to make the `grep`
pattern tolerate color codes where they may be present. But this
would make the pattern considerably longer, more complicated, less
less readable, and easier to get wrong.)