Idea for expanded and clarified `ci-check-msrv` recipe doc
Probably it should not be in quite this form in the `justfile`,
though, because only the `#` line immediately preceding the first
line that is formally part of the recipe is shown in `just --list`.

This only changes the `justfile`. It does not modify the `##`
comment in the corresponding `Makefile` rule.