doc: Fix accessibility bug in MSRV badge
The SVG badge showing the project's MSRV, `etc/msrv-badge.svg`, has
four occurrences of the MSRV in it. Three had the current intended
MSRV. But the top-level `svg` element's `aria-label` attribute
still had the value `rustc: 1.70.0+`, having not been updated the
most recent two times the badge SVG was adjusted for a new MSRV.
That would, at least potentially, result in screen readers and
possibly other software saying/outputting 1.70.0 instead of 1.75.0.

This changes the `aria-label` value to `rustc: 1.75.0+`, as
intended.

This also rephrases the comment in `msrv.yml` about updating the
badge when changing the MSRV, to mention the need to change "all
occurrences".

(We may want to add a CI check to make sure the badge is fully
consistent with the current MSRV, but that isn't attempted here.)