merge-tools builtin: property-based test round-tripping partial changes
This adds another property-based test which tests for a different
property: after applying only some of the changes in a diff, does the
application of the remaining changes result in the "right" tree?

There are some caveats which changes can be safely selected without
breaking this "roundtrip" property, explained in the doc comment.

Inferring the complementary change selection in the format expected by
scm_report is also more tricky than one might naively expect.

The selection is also subject to random generation and is represented by
a `Vec<bool>` that is wrapped around the actual state machine and that
is successively drawn from to decide which change to select. Otherwise,
the presence of this selection mask does not affect the state machine.
I chose `Vec<bool>` other an actual numeric bitmask because proptest
shrinks the vector more sensibly.