Remove unnecessary `jq` validation of `rust-version`
`cargo metadata` already validates the version, checking that it's
a string satisfying one of a small number of recognized formats. So
the only validation `jq` needs to do is to make sure it is really
only appending `.0` to the end if it is of the form `X.Y` for
nonempty numeric `X` and `Y` (with no other characters anywhere).

This shortens the `jq` query considerably, removing the unnecessary
validation and also using regular expression replacement to replace
the empty string that follows the a *full* `X.Y` line with `.0`.