Add test for source file name tracking per hunk
This is a test case for the issue fixed in
7435ed5a9a7370a12332e12bd40fdbc757284a85. The issue was related to the
fact that, before the fix, `source_file_name` was only tracked globally,
as opposed to locally, i. e. on a per-hunk basis.

The test case constructs a repository that has the same rename twice, in
two different branches. Tracking `source_file_name` globally is
insufficient to capture this history which is why this test fails
without the fix.