Get only patch version updates to `imara-diff`
This configures Dependabot version updates to keep `imara-diff` at
0.1.*. It does not affect Dependabot security updates, nor other
ways of updating besides Dependabot, nor other crate dependencies
if they are unrelated to `imara-diff`. See:

- https://github.blog/changelog/2021-05-21-dependabot-version-updates-can-now-ignore-major-minor-patch-releases/
- https://docs.github.com/en/code-security/dependabot/working-with-dependabot/dependabot-options-reference#ignore--

This is by the same technique as used in:
https://github.com/GitoxideLabs/cargo-smart-release/commit/aeb91ee22bcd8aadd64a3261e3de2ed985c99485

See discussion in #2068 for details for why we are not upgrading
`imara-diff` to 0.2 at this time.