Don't run `release.yml` on `push` except in testing
For now, this makes it so the `workflow_dispatch` event has to be
used to run the `release.yml` workflow, except for fake testing
releases (tags that end in `-DO-NOT-USE`).

As noted in the explanatory comment, the `push` event rarely if
ever occurs when tags are pushed to the `GitoxideLabs/gitoxide`
repository. It is uncommon for a `gitoxide` crate release to be
tagged at a commit that does not also have three or more library
crates tagged. But `cargo smart-release` pushes all relevant tags
at once, and GitHub Actions currently does not register a `push`
event when more than three tags are pushed together. So the
`release.yml` workflow is run via `workflow_dispatch` instead.

The preceding commit adds `justfile` recipes to make it easier to
trigger `release.yml` via `workflow_dispatch`. But this runs the
risk that, in the rare case that there are few enough tags pushed
for the `push` trigger to work, the workflow might be run more than
once for the same release. Therefore, this prevents `push` from
ever triggering the workflow for tags representing actual releases.

See #1970 for some other details.

(Allowing `push` to still work in testing makes it easier to test
in a fork without risking accidentally triggering the workflow in
the upstream repository. So the pattern is narrowed to still allow
that, rather than being removed altogether, at least for now.)