Skip more performance tests on Windows ARM
The new pattern `speed` matches the `find_speed` test, and may
match some future performance tests as well.

This is not needed to make tests pass, as `find_speed` was already
passing. However, the `test-fixtures-windows` ARM job takes a very
long time, and this seems that it may be contributing. (It was also
an oversight not to skip it when skipping other performance tests.)

However, the pattern `fuzzed_timeout` is intentionally kept, rather
than being broadened into the pattern `timeout`, since the latter
would match several tests that seem always to complete very fast,
and that test important functionality. (Even `fuzzed_timeout`
itself *usually* completes in less than a tenth of a second.)