Retest only `gix-path`, not `gix-testtools`
`gix-testtools` currently uses a hard-coded `GIT_PROGRAM` value of
`git.exe` on Windows.

`gix-testtools` uses `GIT_PROGRAM` directly in `GIT_CORE_DIR`,
`parse_git_version`, `run_git`, and `populate_meta_dir`, as well as
in its own test case `configure_command_clears_external_config`.

When the path `git.exe` cannot be used to execute the `git`
program, these `gix-testtools` tests fail in `test-fast` on
`ubuntu-latest`:

- tests::bash_program_absolute_or_unrooted
- tests::bash_program_ok_for_platform
- tests::bash_program_unix_path
- tests::configure_command_clears_external_config

Various other functionality of `gix-testtools` is also likely to
fail. `gix-testtools` will have to be fixed to work with this,
probably by causing it to use facilities provided by `gix-path`, as
discussed in #1992 comments, in #1886, and elsewhere.

(Those discussions are with a focus on finding shells through
`git`, rather than finding `git` itself. But using `gix-path`
facilities in `gix-testtools` should be able to solve both, for
when the `git` executable can be found on Windows but not in a
`PATH` search for `git.exe`.)

But let's see if we can already successfully rerun the `gix-path`
tests with `git.exe` not in `PATH`. This removes the step to retest
`gix-testtools`, for now. This is with the idea that if rerunning
`gix-path` tests works in the `test-fast` job on `windows-latest`
and the tests pass there but some fail on `windows-11-arm`, then
they will have surfaced bugs affecting our use of ARM64 builds of
Git for Windows that can be fixed even *before* `gix-testtools` is
modified to use more `gix-path` facilities to find `git` and
associated executables.