Allow empty-valued commit headers in more places
I just fixed a problem where a `gpgsig ` header line in a historical
commit in the git.git repository itself was mistakenly causing a parse
error. The culprit was, of course, that the header value is empty, and
the existing code expected the value _not_ to be empty.

Handle similar issues in other locations, too.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>